Adding new folders and old code (unrefactored)
